+++ packages/version/index.js
@@ -1,5 +1,5 @@
 module.exports = {
   "decorator": "3.17.1",
   "core": "3.17.1",
-  "timestamp": 1725551473848
+  "timestamp": 1726324012264
 };