+++ scripts/rollback/rollback-3.17.3.sh
@@ -1,0 +1,54 @@
+#!/bin/bash
+# timestamp: 1726323927492
+
+npm dist-tag add @midwayjs/axios@3.17.1 latest
+npm dist-tag add @midwayjs/bull@3.17.1 latest
+npm dist-tag add @midwayjs/bull-board@3.17.1 latest
+npm dist-tag add @midwayjs/busboy@3.17.1 latest
+npm dist-tag add @midwayjs/code-dye@3.17.1 latest
+npm dist-tag add @midwayjs/cos@3.17.1 latest
+npm dist-tag add @midwayjs/cross-domain@3.17.1 latest
+npm dist-tag add @midwayjs/grpc@3.17.1 latest
+npm dist-tag add @midwayjs/http-proxy@3.17.1 latest
+npm dist-tag add @midwayjs/i18n@3.17.1 latest
+npm dist-tag add @midwayjs/info@3.17.1 latest
+npm dist-tag add @midwayjs/jwt@3.17.1 latest
+npm dist-tag add @midwayjs/mikro@3.17.2 latest
+npm dist-tag add @midwayjs/mongoose@3.17.2 latest
+npm dist-tag add @midwayjs/passport@3.17.1 latest
+npm dist-tag add @midwayjs/security@3.17.1 latest
+npm dist-tag add @midwayjs/static-file@3.17.1 latest
+npm dist-tag add @midwayjs/swagger@3.17.2 latest
+npm dist-tag add @midwayjs/tags@3.17.1 latest
+npm dist-tag add @midwayjs/typegoose@3.17.2 latest
+npm dist-tag add @midwayjs/upload@3.17.1 latest
+npm dist-tag add @midwayjs/validate@3.17.1 latest
+npm dist-tag add @midwayjs/version@3.17.2 latest
+npm dist-tag add @midwayjs/express@3.17.1 latest
+
+# Changes:
+
+#  - @midwayjs/axios: 3.17.1 => 3.17.3
+#  - @midwayjs/bull: 3.17.1 => 3.17.3
+#  - @midwayjs/bull-board: 3.17.1 => 3.17.3
+#  - @midwayjs/busboy: 3.17.1 => 3.17.3
+#  - @midwayjs/code-dye: 3.17.1 => 3.17.3
+#  - @midwayjs/cos: 3.17.1 => 3.17.3
+#  - @midwayjs/cross-domain: 3.17.1 => 3.17.3
+#  - @midwayjs/grpc: 3.17.1 => 3.17.3
+#  - @midwayjs/http-proxy: 3.17.1 => 3.17.3
+#  - @midwayjs/i18n: 3.17.1 => 3.17.3
+#  - @midwayjs/info: 3.17.1 => 3.17.3
+#  - @midwayjs/jwt: 3.17.1 => 3.17.3
+#  - @midwayjs/mikro: 3.17.2 => 3.17.3
+#  - @midwayjs/mongoose: 3.17.2 => 3.17.3
+#  - @midwayjs/passport: 3.17.1 => 3.17.3
+#  - @midwayjs/security: 3.17.1 => 3.17.3
+#  - @midwayjs/static-file: 3.17.1 => 3.17.3
+#  - @midwayjs/swagger: 3.17.2 => 3.17.3
+#  - @midwayjs/tags: 3.17.1 => 3.17.3
+#  - @midwayjs/typegoose: 3.17.2 => 3.17.3
+#  - @midwayjs/upload: 3.17.1 => 3.17.3
+#  - @midwayjs/validate: 3.17.1 => 3.17.3
+#  - @midwayjs/version: 3.17.2 => 3.17.3
+#  - @midwayjs/express: 3.17.1 => 3.17.3